# backend/swagger/swagger_dxf_routes.yaml
openapi: 3.0.3
info:
  title: DXF文件服务 API
  description: DXF文件处理、样式分析、Martin服务发布等功能的API接口
  version: 1.0.0
  contact:
    name: SHP Service Team
    email: support@shpservice.com
servers:
  - url: /api/dxf
    description: DXF文件服务
security:
  - BearerAuth: []
paths:
...
  /martin-services/{service_id}:
    parameters:
      - name: service_id
        in: path
        required: true
        schema:
          type: integer
        description: 服务ID
        example: 1234567890123456789
    
    delete:
      tags:
        - Dxf文件相关
      summary: 删除DXF Martin服务
      description: 删除指定的DXF Martin服务，同时清理PostGIS表
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "DXF Martin服务已删除"
                  deleted_service:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "1234567890123456789"
                      table_name:
                        type: string
                        example: "vector_a1b2c3d4"
                      original_filename:
                        type: string
                        example: "drawing.dxf"
        '404':
          description: 服务不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "服务不存在"
        '500':
          description: 删除失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "删除服务失败: 数据库错误"
...
  /style-templates:
    get:
      tags:
        - Dxf文件相关
      summary: 获取DXF样式模板
      description: 获取预定义的DXF样式模板列表
      responses:
        '200':
          description: 获取样式模板成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  templates:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "template_1"
                        name:
                          type: string
                          example: "建筑图纸模板"
                        description:
                          type: string
                          example: "适用于建筑CAD图纸的样式模板"
                        layers:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                                example: "WALLS"
                              color:
                                type: string
                                example: "#FF0000"
                              lineweight:
                                type: number
                                example: 0.5
                              linetype:
                                type: string
                                example: "CONTINUOUS"
                        colors:
                          type: object
                          properties:
                            "1":
                              type: string
                              example: "#FF0000"
                            "2":
                              type: string
                              example: "#00FF00"
                            "3":
                              type: string
                              example: "#0000FF"
                        created_at:
                          type: string
                          format: date-time
                          example: "2024-01-15T10:30:00Z"
        '500':
          description: 获取样式模板失败
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "获取样式模板失败: 数据库连接错误"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证令牌
...